fix: move default app rules to constant
